inference/requirements.txt
numpy==1.18.1
pillow==8.2.0